% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{adeconv}
\alias{adeconv}
\title{Absorption band deconvolution}
\usage{
adeconv(spec, sr, start, lower, upper, ef, model = c("Gauss", "Lorentz",
  "Voigt"), maxit = 50000, reltol = 1e-12)
}
\arguments{
\item{spec}{A vector with the spectra to be deconvolved.}

\item{sr}{A vector with the spectral reference of the spectra.}

\item{start}{A matrix with first guess parameters. See details.}

\item{lower}{A matrix with the lower limit of the parameters to be fitted.
See details.}

\item{upper}{A matrix with the upper limit of the parameters to be fitted.
See details.}

\item{ef}{Error function to be minimized.}

\item{model}{One of 'Gauss', 'Lorentz' or 'Voigt'. See details.}

\item{maxit}{Maximum number of iterations.}

\item{reltol}{Relative tolerance for convergence criteria.}
}
\value{
The return of optim, with fitted coefficients converted to matrix
format for convenience.
}
\description{
This function deconvolves a measured signal into a sum of model functions.
}
\details{
This function is a wrapper for the \code{optim} function, with the 
Nelder-Mead method, for deconvolution of a spectrum as sum of Gaussian, 
Lorentzian or Voigt curves. It is appropriate for deconvolution of absorption 
peaks if the peaks are present in or just beyond the limits of the 
measured spectrum. When applied for deconvolution of pigment absorption, it 
is important that other absorption signals have been removed form the 
spectrum for a meaningful result. The spectral reference scale may be any 
desired (e.g., energy, wavenumber, wavelength), but the fitting space, 
\code{start} values and \code{sr} must be specified in the correct scale.

The arguments \code{start}, \code{lower}, and \code{upper}, must be three 
column matrices for models 'Gauss' and 'Lorentz', with center, sigma 
('Gauss') or hwhm ('Lorentz') and scale, respectively and peaks per
row. For the 'Voigt' model, it should have four columns, with center, sigma, 
hwhm and scale (in this order). The number of absorption bands used depends 
on the number of rows and the fitting space is completely defined by the 
lower and upper boundaries (may be infinite). The boundaries are enforced by 
returning NA on the error function evaluation when optim test parameter 
values outside the fitting space.

The spectra to be deconvolved may contain NA, but not \code{sr}, 
\code{start}, \code{lower}, and \code{upper}. \code{maxit} and \code{reltol}
are passed to optim. A set of error functions is provided: \code{mard}, 
\code{rmsd}, \code{mlrt}, and \code{rrsd}. For more details on the Gaussian, 
Lorentzian and Voigt parameters, see \code{apeaks}.
}
\references{
Hoepffner, N.; Sathyendranath, S. 1991. Effect of pigment composition on 
absorption properties of phytoplankton. Marine Ecology Progress Series 73, 
11-23. DOI: 10.3354/meps073011
}
\seealso{
\code{\link{apeaks}}, \code{\link{mard}}, \code{\link{rmsd}}, 
\code{\link{mlrt}}, and \code{\link{rrsd}}
}
