% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{apeaks}
\alias{apeaks}
\title{Calculate absorption spectra}
\usage{
apeaks(center, sigma, hwhm, scale, sr, model = c("Gauss", "Lorentz",
  "Voigt"), peaks = FALSE)
}
\arguments{
\item{center}{Center position reference of each absorption band.}

\item{sigma}{Standard deviation for the Gaussian model.}

\item{hwhm}{Half width at half maximum for the Lorentzian model.}

\item{scale}{Scale height at the center of each absorption band.}

\item{sr}{Position reference at which the absorption models should be 
evaluated in any suitable scale (energy, frequency, wavelength).}

\item{model}{One of 'Gauss', 'Lorentz' or 'Voigt'. See details.}

\item{peaks}{Logical. Should the peak curves for each absorption band be 
returned?}
}
\value{
If peaks is FALSE, a numeric vector with the summed peaks. If peaks 
is true, a matrix is returned, with peaks per row.
}
\description{
This function evaluates an absorption band model at the desired reference 
positions.
}
\details{
Observed absorption bands are not sharp due to several line 
broadening process grouped into homogeneous processes represented by a 
Lorentzian profile and inhomogeneous processes represented by a Gaussian 
profile. The Voigt profile combine both processes. 

The line width parameters for each profile are requested in their most common
formats, standard deviation for the Gaussian model and half width at half 
maximum (HWHM) for the Lorentzian model. The HWHM is related to the standard 
deviation for the Gaussian function by:
HWHM = sqrt(2*log(2)) * sigma

The Voigt profile is calculated with the analytical approximation of Liu 
et al. (2001) and the empirical approximation of Olivero & Longbothum (1977) 
for the HWHM of the Voigt profile.

Units for the parameters depend on the scale used to represent the absorption 
process (e.g., energy, frequency, wavenumber, wavelength).
}
\examples{
sr <- seq(-10, 10, 0.01)
pg <- apeaks(0, 1, 0, 1, sr = sr, 'Gauss')
pl <- apeaks(0, 0, sqrt(2*log(2)), 1, sr = sr, 'Lorentz')
pv <- apeaks(0, 0.7638, 0.4 * sqrt(2*log(2)), 1, sr = sr, 'Voigt')
plot(sr, pg, type = 'l', col = 2, ylab = 'Height')
lines(sr, pl, col = 3)
lines(sr, pv, col = 4)
legend('topright', c('Gauss', 'Lorentz', 'Voigt'), col = 2:4, lty = 1, bty = 'n')
legend('topright', expression(HWHM == sqrt(2*log(2))), bty = 'n')
}
\references{
Liu, Y.; Lin, J.; Huang, G.; Guo, Y.; Duan, C. 2001. Simple empirical 
analytical approximation to the Voigt profile. Journal of the Optical Society 
of America B 18, 5, 666-672. DOI: 10.1364/JOSAB.18.000666

Olivero, J. J.; Longbothum, R. L. 1977. Empirical fits to the Voigt line 
width: A brief review. Journal of Quantitative Spectroscopy and Radiative 
Transfer 17, 2, 233-236. DOI: 10.1016/0022-4073(77)90161-3
}
\seealso{
\code{\link{adeconv}}
}
