% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avw.R
\name{avw}
\alias{avw}
\title{Calculate the Apparent Visible Wavelength (AVW)}
\usage{
avw(wave, refl, sensor)
}
\arguments{
\item{wave}{Reference centre wavelength (nm).}

\item{refl}{Reflectance data (1/sr or unitless).}

\item{sensor}{One of: 'hyperspectral' or 'aeronet-oc'. See details.}
}
\description{
The function calculates the Apparent Visible Wavelength (AVW) for hyperspectral 
and selected multispectral sensors.
}
\details{
The Apparent Visible Wavelength (AVW) is the weighted harmonic mean
wavelength, with the weights given by the reflectance for all wavelengths in
the visible range (400 nm to 700 nm).

For multispectral sensors, the calculated AVW can be concerted to the hyperspectral
equivalent AVW by using a regression function for a given multispectral sensor waveband 
set using simulated hyperspectral data. Currently the only multispectral sensor supported
is the AERONET-OC. The base set of wavebands should be used as input (412, 443, 488, 547, 667).
}
\references{
Vandermeulen, R.; Mannino, A.; Craig, S.; Werdell, P.G. 2020. 150 shades of green: Using the full 
spectrum of remote sensing reflectance to elucidate color shifts in the ocean. Remote Sensing of 
Environment 247, 111900. DOI: https://doi.org/10.1016/j.rse.2020.111900
}
