% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{BeginBuiltinCapture}
\alias{BeginBuiltinCapture}
\title{Setup capture of builtin functions

Sets up capturing of builtin functions}
\usage{
BeginBuiltinCapture(internal = FALSE, functions = builtins(internal),
  indexes, verbose = testrOptions("verbose"))
}
\arguments{
\item{internal}{wheather only internals should be captured, or all builtins}

\item{functions}{list of functions to be decorate}

\item{indexes}{specific indexes from functions vector}

\item{verbose}{if to output debug information}
}
\description{
Setup capture of builtin functions

Sets up capturing of builtin functions
}
\seealso{
SetupCapture
}

