% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhcoat.R
\name{bhcoat}
\alias{bhcoat}
\title{Calculate the efficiency factors for coated spherical particles}
\usage{
bhcoat(ri, ro, ni, no, lambda)
}
\arguments{
\item{ri}{Radius of the inner sphere.}

\item{ro}{Radius of the outer sphere.}

\item{ni}{Refractive index of the inner sphere relative to surrounding 
medium.}

\item{no}{Refractive index of the outer sphere relative to surrounding 
medium.}

\item{lambda}{Wavelength in the surrounding medium, in the same units as the 
radius.}
}
\value{
A numerical vector with the efficiency factors for extinction, 
absorption, scattering and backscattering.
}
\description{
This function calculates the efficiency factors of extinction, absorption, 
scattering and backscattering for particle consisting of two concentric 
spheres.
}
\examples{
ni  <-  complex(real = 1.02, imaginary = 0.006)
no  <-  complex(real = 1.10, imaginary = 0.0)
bhcoat(ri = 5, ro = 5.2, ni = ni, no = no, lambda = 0.675)


}
\references{
Bohren, C. F.; Huffman, D. R. 1983. Absorption and Scattering of Light by 
Small Particles. John Wiley & Sons, New York, 533 pp.
}
