% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/water_density.R
\name{d_water}
\alias{d_water}
\title{Density of water}
\usage{
d_water(Tc = 20, S = 0, P = 0)
}
\arguments{
\item{Tc}{Temperature (ºC).}

\item{S}{Salinity (parts per thousand).}

\item{P}{Pressure (bar in excess of 1 ATM).}
}
\description{
This function calculates the water density as a function of pressure, 
temprature and salinity.
}
\details{
The density of pure (saline) water at diferent pressures is based on 
the UNESCO (1981) report, as recomended by Zhang et al. (2009) and Libes 
(2009). The UNESCO equation has a relative error of 0.0004% (Zhang et al.,
2009). The actual function was implemented with corrections by from 
Forfonoff (1985), as recomended by Libes (2009).
}
\examples{
# Average seawater density at surface under 1 ATM:
d_water(Tc = 3.5, S = 34.72, P = 0)

}
\references{
Fofonoff, N. P. 1985. Physical properties of seawater: A new salinity scale 
and equation of state for seawater. Journal of Geophysical Research: Oceans 
90, C2, 3332-3342. DOI: 10.1029/JC090iC02p03332

Libes, S. 2009. Introduction to Marine Biogeochemistry. 2nd Edition. Academic 
Press, 928 pp.

Zhang, X. Hu, L. 2009. Estimating scattering of pure water from density 
fluctuation of the refractive index. Optics Express 17, 3, 1671-1678. DOI: 
10.1364/OE.17.001671
}
