% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iop_water.R
\name{.beta_water_90}
\alias{.beta_water_90}
\title{Volume scattering at 90 degrees (m^-1 sr^-1)}
\usage{
.beta_water_90(lambda, S = 0, Tc = 20, P = 0)
}
\description{
Volume scattering at 90 degrees (m^-1 sr^-1)
}
