% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_args.R
\name{.check_vect_args}
\alias{.check_vect_args}
\title{Will return TRUE if the condition is NOT met. Condition is that the lengths 
of all vector arguments must be an iteger multiple of the the length of the
longer vector.}
\usage{
.check_vect_args(arg_ls)
}
\description{
Will return TRUE if the condition is NOT met. Condition is that the lengths 
of all vector arguments must be an iteger multiple of the the length of the
longer vector.
}
