% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iop_water.R
\name{.dlna0dS}
\alias{.dlna0dS}
\title{Partial derivative of the natural logarithm of solvent activity with respect 
to salinity, dln(a0)/dS is from Millero & Leung (1976). Has an estimated 
error of 0.04%.}
\usage{
.dlna0dS(S, Tc)
}
\description{
Partial derivative of the natural logarithm of solvent activity with respect 
to salinity, dln(a0)/dS is from Millero & Leung (1976). Has an estimated 
error of 0.04%.
}
