% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iop_water.R
\name{.dnds}
\alias{.dnds}
\title{Partial derivative of seawater refractive index with respect to salinity, 
dn/dS. Based on Quan & Fry (1995).}
\usage{
.dnds(lambda, Tc)
}
\description{
Partial derivative of seawater refractive index with respect to salinity, 
dn/dS. Based on Quan & Fry (1995).
}
