% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fastr-translator.R
\name{FastrTagIgnored}
\alias{FastrTagIgnored}
\title{Function that tags failing test cases for FastR as ignored

This function tags test cases with Ignore preporcessor. It requires a file with results of running tests}
\usage{
FastrTagIgnored(fastr.test.root, result.file)
}
\arguments{
\item{fastr.test.root}{folder with FastR test cases}

\item{result.file}{file with result of running FastR test cases}
}
\description{
Function that tags failing test cases for FastR as ignored

This function tags test cases with Ignore preporcessor. It requires a file with results of running tests
}

