% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testgen.r
\name{GenerateTC}
\alias{GenerateTC}
\title{Generates a testcase for closure function}
\usage{
GenerateTC(symb, vsym, func, body, argv, warn, retv, errs,
  use.get.anywhere = TRUE)
}
\arguments{
\item{symb}{symbols to be initialized before the call}

\item{vsym}{values of the symbols}

\item{func}{closure function name}

\item{body}{closure body}

\item{argv}{input arguments for a closure function call}

\item{warn}{expected warnings}

\item{retv}{expected return value}

\item{use.get.anywhere}{whether to try to get body of the function in the loaded namespaces}

\item{retv}{expected errors}
}
\description{
This function generates a test case for builtin function using supplied arguments. All elements should be given as text.
}
\seealso{
TestGen ProcessClosure
}

