% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tc-db.R
\name{GetAllFiles}
\alias{GetAllFiles}
\title{Get all files with specific pattern

This function is respinsible for leturning all files from specified folder}
\usage{
GetAllFiles(root, pattern = ".[rR]$", full.names = T)
}
\arguments{
\item{root}{input folder}

\item{pattern}{pattern of files to be searched for}

\item{full.names}{if full path to files should be returned}
}
\description{
Get all files with specific pattern

This function is respinsible for leturning all files from specified folder
}

