% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconv.R
\name{gpeaks}
\alias{gpeaks}
\title{Calculate Gaussian absorption spectra}
\usage{
gpeaks(center, sigma, scale, sr, peaks = FALSE)
}
\arguments{
\item{center}{Center spectral reference of each Gaussian curve. 
Alternatively, a matrix in which columns are center, sigma and
scale. See details.}

\item{sigma}{Standard deviation in the appropriate spectral reference units.}

\item{scale}{Scale height at the center of each Gaussian curve.}

\item{sr}{Spectral reference at which Gaussians should be evaluated.}

\item{peaks}{Logical. Should the Gaussian curves for each absorption band be 
returned?}
}
\value{
If peaks is FALSE, a numeric vector with the summed peaks. If peaks 
is true, a matrix is returned, with peaks per row.
}
\description{
This function evaluates Gaussian function with the input parameters at the 
desired spectral reference points.
}
\details{
Units for center and sigma depend on the spectral scale used during
parameter fit and may be any convenient scale were absorption bands can be 
described by a Gaussian curve (e.g., wavelength in nm or wavenumber in 1/cm).

If \code{center} is matrix, it must have three columns (center, sigma, 
scale), with absorption peaks by row. Note that if sigma and scale parameters
take precedence and will be used instead of the matrix columns if specified.
}
\seealso{
\code{\link{gdeconv}}
}
