% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testr.r
\name{is.generatorContentFunction}
\alias{is.generatorContentFunction}
\title{Checks that the given function has a proper signature for a generator content function.

A generator content function takes the index of the value to be returned that must be named i, the generator itself named g, and the environment of the current test called env. The environment is a list containing values of all previously defined generators in the particular test instance.
The function returns true if the given function has three arguments, named i, g and env. False otherwise.}
\usage{
is.generatorContentFunction(f)
}
\arguments{
\item{f}{function whose signature is to be checked.}
}
\value{
TRUE if the function has the signature of a generator content function (arguments i, g and env) and FALSE otherwise
}
\description{
Checks that the given function has a proper signature for a generator content function.

A generator content function takes the index of the value to be returned that must be named i, the generator itself named g, and the environment of the current test called env. The environment is a list containing values of all previously defined generators in the particular test instance.
The function returns true if the given function has three arguments, named i, g and env. False otherwise.
}
\examples{
is.generatorContentFunction(function(i,g,env) { g$expr[[i]]} )
}
\seealso{
customGenerator
}

