% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refractive_indexes.R
\name{liquid_DFRI}
\alias{liquid_DFRI}
\title{Calculate the density fluctuation of the refractive index (DFRI) for liquids}
\usage{
liquid_DFRI(n, d)
}
\arguments{
\item{n}{Refractive index of the medium (unitless).}

\item{d}{Density of the medium (kg dm^-3).}
}
\value{
A numeric vector with the partial derivative of the squared 
refractive index with respect to density (dn^2 / dd, dm^3 kg^-1), with 
density given in kg dm^-3.
}
\description{
This function calculates the density fluctuation of the refractive index 
(DFRI) for liquids. This is the partial derivative of the squared refractive 
index with respect to density when keeping temperature (isothermic) or 
pressure (isobaric) constant.
}
\details{
Between the different formulations for the DFRI, the formulation of 
Proutiere et al. (1992) show the best experimental comparison with observed 
data of scattering, when deriving scattering by pure water from density 
fluctuations (Zhang et al., 2009). Strictly, the formulation was derived for 
isothermal conditions, but theory supports that it should be the same as 
isobaric (at least approximately).
}
\examples{
# Water entry on Table II of Proutiere et al. (1992):
liquid_DFRI(1.3330, 0.9982)

}
\references{
Proutiere, A.; Megnassan, E.; Hucteau, H. 1992. Refractive index and density 
variations in pure liquids: A new theoretical relation. The Journal of 
Physical Chemistry 96, 3485-3489. DOI: 10.1021/j100187a058

Zhang, X. Hu, L. 2009. Estimating scattering of pure water from density 
fluctuation of the refractive index. Optics Express 17, 3, 1671-1678. DOI: 
10.1364/OE.17.001671
}
\seealso{
\code{\link{n_water}}
}
