% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorfn.R
\name{mlrt}
\alias{mlrt}
\alias{mard}
\alias{rmsd}
\alias{rrsd}
\title{Error functions for minimization}
\usage{
mlrt(x, y, w = rep(1, length(x)))

mard(x, y, w = rep(1, length(x)))

rmsd(x, y, w = rep(1, length(x)))

rrsd(x, y, w = rep(1, length(x)))
}
\arguments{
\item{x}{Reference vector.}

\item{y}{Test vector.}

\item{w}{Weights vector.}
}
\value{
A scalar with the difference between the two vectors.
}
\description{
Those functions calculate the (weighted) difference between two vectors.
}
\details{
Those functions are commonly used in the literature as error 
functions for minimization in inversion problems. \code{mlrt} computes the
mean log ratio, \code{mard} computes the mean absolute relative difference, 
\code{rmsd} computes the root mean squared difference, and \code{rrse} 
computes the relative root squared error between two vectors. All function 
accept a vector of weights, as in \code{weighted.mean}.
}
\seealso{
\code{\link{weighted.mean}}
}
