% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refractive_indexes.R
\name{n_calcite}
\alias{n_calcite}
\title{Dispersion formula for the refractive index of calcite}
\usage{
n_calcite(lambda, comp = F)
}
\arguments{
\item{lambda}{Wavelength in vacuum (nm).}

\item{comp}{Logical. Should the ordinary and extraordinary indexes be returned?}
}
\value{
A numeric vector with the refractive index (unitless) of calcite 
relative to vaccum.
}
\description{
This function calculates the ordinary and extraordinary refractive index of 
calcite relative to vaccum. See details.
}
\details{
Calcite is birefringent with two orientation dependent indexes of 
refraction, the extraordinary index of refraction (propagation along the 
optical axis) and the ordinary index of refraction (propagation orthogonal to 
optical axis). Values are provided at "room temperature".

If "comp" is FALSE (default), the average refractive index is returned. 
Otherwise a matrix is returned.
}
\examples{
# Retrieve the average real part of the refractive index of calcite in the 
visible range:
n_calcite(lambda = 400:700)

}
\references{
Ballard, S. S.; Browder, J. S.; Ebersole, J. F. 1982. Index of Refraction. 
In: Gray D. E. Ed., American Institute of Physics Handbook, McGraw-Hill, New 
York.

Fournier, G. R.; Ardouin, J.-P.; Levesque, M. 2018. Modeling Sea Bottom 
Hyperspectral Reflectance. Applied Sciences 8, 12, 23 p. DOI: 
10.3390/app8122680

Ghosh, G. 1999. Dispersion-equation coefficients for the refractive index and 
birefringence of calcite and quartz crystals. Opt. Commun. 163, 95–102.
}
\seealso{
\code{\link{n_air}}, \code{\link{n_water}}, \code{\link{n_quartz}}, 
\code{\link{n_cellulose}}, \code{\link{snell}}, \code{\link{fresnel}}
}
