% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refractive_indexes.R
\name{n_cellulose}
\alias{n_cellulose}
\title{Dispersion formula for the refractive index of cellulose}
\usage{
n_cellulose(lambda)
}
\arguments{
\item{lambda}{Wavelength in vacuum (nm).}
}
\value{
A numeric vector with the refractive index (unitless) of cellulose 
relative to vaccum.
}
\description{
This function calculates the real part of the refractive index of cellulose.
}
\examples{
# Retrieve the real part of the refractive index of cellulose in the visible
# range:
n_cellulose(lambda = 400:700)

}
\references{
Ballard, S. S.; Browder, J. S.; Ebersole, J. F. 1982. Index of Refraction. 
In: Gray D. E. Ed., American Institute of Physics Handbook, McGraw-Hill, New 
York.

Fournier, G. R.; Ardouin, J.-P.; Levesque, M. 2018. Modeling Sea Bottom 
Hyperspectral Reflectance. Applied Sciences 8, 12, 23 p. DOI: 
10.3390/app8122680

Sultanova, N.; Kasarova, S.; Nikolov, I. 2009. Dispersion properties of 
optical polymers. Acta Phys. Pol. A 116, 585–587. DOI:
}
\seealso{
\code{\link{n_air}}, \code{\link{n_water}}, \code{\link{n_calcite}}, 
\code{\link{n_quartz}}, \code{\link{snell}}, \code{\link{fresnel}}
}
