% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refractive_indexes.R
\name{n_water}
\alias{n_water}
\title{Dispersion formula for the refractive index of liquid water}
\usage{
n_water(lambda, Tc, S, P, type = c("real", "complex"))
}
\arguments{
\item{lambda}{Wavelength in vacuum [200,1100] (nm).}

\item{Tc}{Temperature [0,30] (ºC).}

\item{S}{Salinity [0,40] (parts per thousand).}

\item{P}{Pressure [0,1028] (bar in excess of 1 ATM).}

\item{type}{One of: "real" or "complex". See details.}
}
\value{
A numeric (or complex) vector of the refractive index (unitless) of 
pure (saline) water relative to vacuum .
}
\description{
This function calculates the real part of the refractive index of pure water
relative to vaccum. The complex refractive index can also be returned. See 
details.
}
\details{
The temperature and salinity dependent spectral refractive index of 
(saline) water is based on the empirical dispersion equation parametrized by 
Quan & Fry (1995) for the visible range. Huibers (1997) analyzed the 
available data and models and attested that the model could be extrapolated 
to the UV and NIR (220 to 1100 nm).

The Quan & Fry (1995) formulation is for standard atmosphere pressure. It can 
be converted to any desired pressure with the isothermic density derivative 
of the refractive index of liquids parametrized by Proutiere et al. (1992). 
That pressure, in bar, should be in excess of surface pressure, i.e., to 
retrieve surface values, P = 0.

The imaginary part of the refractive index of pure water is provided here as 
independent of temperature, salinity and pressure, taken from the tabulated 
values provided by Segeistein (1975). Note that in the UV range it is 
expected that the absorption and therefore the imaginary part of the 
refractive index be dependent of dissolved salts and dissolved oxygen (Jonaz 
& Fournier, 2007).

The original parametrization of Quan & Fry (1995) is based on refractive 
index relative to air measured by Austin & Halikas (1976), and the function 
will return the values relative to vacuum by using the refractive index of 
standard dry air (\code{n_air}), as recommended by Zhang et al. (2009).

Another parametrization with comparable to that of Quan & Fry (1995) in the 
visible range Huibers, 1997) is that of Schiebener et al. (1990). This model
has an extended range (200 to 2500 nm) and could be added as an option in a
future version.
}
\examples{
# Real part of the refractive index of average seawater at 1 atmospheric 
# pressure and in the visible range:
n_water(lambda = 400:700, S = 34.72, Tc = 3.5, P = 0)

}
\references{
Austin, R. W.; Halikas, G. 1976. The index of refraction of seawater. SIO 
Ref. 76-1, Scripps Institution of Oceanography, La Jolla, California.

Huibers, P. D. T. 1997. Models for the wavelength dependence of the index of 
refraction of water. Applied Optics 36, 16, 3785-3787. DOI: 
10.1364/ao.36.003785

Jonasz, M.; Fournier, G. R. 2007. Light Scattering by Particles in Water - 
Theoretical and Experimental Foundations. Academic Press, 715 pp.

Proutiere, A.; Megnassan, E.; Hucteau, H. 1992. Refractive index and density 
variations in pure liquids: A new theoretical relation. The Journal of 
Physical Chemistry 96, 3485-3489. DOI: 10.1021/j100187a058

Quan, X.; Fry, E. S. 1995. Empirical equation for the index of refraction of 
seawater. Applied Optics 34, 3477–3480. DOI: 10.1364/AO.34.003477

Schiebener, P.; Straub, J.; Levelt Sengers, J. M. H.; Gallagher, J. S. 1990. 
Refractive index of water and steam as a function of wavelength, temperature 
and density. Journal of Physical and Chemical Reference Data 19, 677–717. 
DOI: 10.1063/1.555859

Segeistein, D. J. 1975. The complex refractive index of water. Master Thesis,
University of Missouri-Kansas City, 175 pp.

Zhang, X. Hu, L. 2009. Estimating scattering of pure water from density 
fluctuation of the refractive index. Optics Express 17, 3, 1671-1678. DOI: 
10.1364/OE.17.001671
}
\seealso{
\code{\link{n_air}}, \code{\link{n_cellulose}}, 
\code{\link{n_calcite}}, \code{\link{n_quartz}}, \code{\link{snell}}, 
\code{\link{fresnel}}
}
