% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/target.r
\name{PrintTest}
\alias{PrintTest}
\title{Prettyprints the test, is intended to be used only internally.}
\usage{
PrintTest(test, code = NULL)
}
\description{
Prettyprints the test, is intended to be used only internally.
}

