% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm_semi_analytical.R
\name{propagate_r}
\alias{propagate_r}
\title{Propagates reflectance across a boundary:
HOW TO GENERALIZE THIS FUNCTION??? MAKE BOTH WAYS, for Rrs and R}
\usage{
propagate_r(rrs)
}
\description{
Propagates reflectance across a boundary:
HOW TO GENERALIZE THIS FUNCTION??? MAKE BOTH WAYS, for Rrs and R
}
