% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwip.R
\name{qwip_qc}
\alias{qwip_qc}
\title{Calculate the QWIP quality control flag}
\usage{
qwip_qc(wave, refl, sensor, score_lim = 0.2)
}
\arguments{
\item{wave}{Reference centre wavelength (nm).}

\item{refl}{Reflectance data (1/sr or unitless).}

\item{sensor}{One of: 'hyperspectral' or 'aeronet-oc'. See details.}

\item{score_lim}{A magnitude of maximum QWIP score to accept in the quality control.}
}
\description{
The function performs the quality control of reflectance spectra using the QWIP approach.
}
\details{
The QWIP quality control calculates the QWIP score and evaluate if it is within the user
defined limits. It also evaluates if the AVW occurs is the expected range for the classified 
water type using a simple three types classification.

For multispectral sensors, the calculated AVW can be concerted to the hyperspectral
equivalent AVW by using a regression function for a given multispectral sensor waveband 
set using simulated hyperspectral data. Currently the only multispectral sensor supported
is the AERONET-OC. The base set of wavebands should be used as input (412, 443, 488, 547, 667).
}
\references{
Dierssen, H.; Vandermeulen, R.; Barnes, B.; Castagna, A.; Knaeps, E.; Vanhellemont, Q. 2022. 
QWIP: A Quantitative Metric for Quality Control of Aquatic Reflectance Spectral Shape Using the 
Apparent Visible Wavelength. Frontiers in Remote Sensing 3. DOI: 
https://doi.org/10.3389/frsen.2022.869611
}
