% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qwip.R
\name{qwip_score}
\alias{qwip_score}
\title{Calculate the QWIP score}
\usage{
qwip_score(wave, refl, sensor)
}
\arguments{
\item{wave}{Reference centre wavelength (nm).}

\item{refl}{Reflectance data (1/sr or unitless).}

\item{sensor}{One of: 'hyperspectral' or 'aeronet-oc'. See details.}
}
\description{
The function calculates the Quality Water Index Polynomial (QWIP) for hyperspectral 
and selected multispectral sensors.
}
\details{
The QWIP score the difference between the observed reflectance Normalized Difference 
Index at 492 nm and 667 nm and the predicted NDI based on an empirical relation against the 
Apparent Visible Wavelength (AVW).

For multispectral sensors, the calculated AVW can be concerted to the hyperspectral
equivalent AVW by using a regression function for a given multispectral sensor waveband 
set using simulated hyperspectral data. Currently the only multispectral sensor supported
is the AERONET-OC. The base set of wavebands should be used as input (412, 443, 488, 547, 667).
}
\references{
Dierssen, H.; Vandermeulen, R.; Barnes, B.; Castagna, A.; Knaeps, E.; Vanhellemont, Q. 2022. 
QWIP: A Quantitative Metric for Quality Control of Aquatic Reflectance Spectral Shape Using the 
Apparent Visible Wavelength. Frontiers in Remote Sensing 3. DOI: 
https://doi.org/10.3389/frsen.2022.869611
}
