% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry_fun.R
\name{rad}
\alias{rad}
\alias{deg}
\title{Convert between radians and degrees}
\usage{
rad(x)

deg(x)
}
\arguments{
\item{x}{Angle (radians or degrees)}
}
\value{
A numeric vector
}
\description{
This functions convert between angles represented in radians and degrees.
}
\seealso{
\code{\link{snell}}, \code{\link{snell_decomp}}
}
