% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coverage.r
\name{ResetCoverageInfo}
\alias{ResetCoverageInfo}
\title{Reset coverage information}
\usage{
ResetCoverageInfo(root)
}
\arguments{
\item{root}{a directory or a single C file that contains or is instrumented VM source.}
}
\description{
This function deletes gcov information files (*.gcda) thus clearing any previously collected coverage information is erased.
}

