% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rsrf}
\alias{rsrf}
\title{Relative Spectral Response Functions}
\format{
An object of class \code{list} of length 5.
}
\usage{
rsrf
}
\description{
The relative spectral response function (RSRF) of several sensors is 
provided. The RSRF are normalized for a maximum of 1 at each band and 
provided as a matrix organized in lists. The first column is the wavelength
reference.
}
\details{
Landsat 8   - RSRF for the Operational Land Imager (OLI) covering in and out-
              of-band response, average over focal plane modules. In band and 
              out-of-band responses were merged by extending the in band 
              response at a constant of 0.001 and keeping the minimum value 
              over the overlap regions. Original data retrieved from USGS in
              June 2019.
Sentinel-2  - RSRF for the Multispectral Instrument (MSI) covering in band 
              response only, average over detector modules. Data retrieved 
              from ESA in June 2019. 
Pleiades    - RSRF for sensors PHRA and PHRB, covering in band response only.
              Retrived from AIRBUS in June 2019.
RapidEye    - RSRF for the Jena Spaceborne Scanner (JSS) covering in band 
              response only. A single RSRF is provided, representative 
              (average?) of the different JSS sensors. Data retrieved from 
              PlanetLabs in June 2019.
PlanetScope - RSRF for sensor classes 0c, 0d, 0e, 0f and 10, including in 
              band response only. A single RSRF is provided, representative 
              (average?) of the different sensors in each class. Classes 0c 
              and 0d have the same RSRF, the same is true for classes 0f and 
              10. 
AERONET-OC  - RSRF of AERONET-OC sensors, including only nominal rectangular
              10 nm wide SRF.
}
\keyword{datasets}
