% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rstat}
\alias{rstat}
\alias{lstat}
\title{Calculate retrieval statistics}
\usage{
rstat(x, y, units = "")

lstat(stats, digits = 3)
}
\arguments{
\item{x}{Reference data.}

\item{y}{Estimated data.}

\item{units}{Units of the data.}

\item{stats}{object created by \code{rstat}.}

\item{digits}{number of digits to be passed to \code{round}.}
}
\value{
A named list with component statistics and units for \code{rstat} and 
a vector of expressions for plot legend.
}
\description{
The function calculates useful descriptive statistics to evaluate the 
inversions and create a vector of strings for plotting.
}
\details{
Those functions are wrappers for descriptive statistics and 
convenient text strings for addition to a plot. They provide the statistics 
of the agreement of the estimated data against the reference data; i.e.
evaluates estimations against the 1:1 relation. Statistics include the mean 
absolute percentage error (MAPE), the mean percentage error (bias), the root 
mean squared error (RMSE), the coefficient of determination (R^2), the number
of data points, and the range of the reference data.

The parameter units is also available for convenience of use with the
\code{lstats} function, to add those statistics to a plot. It should be a 
character string that can be used in expressions. See examples.
}
\examples{
x <- 1:10
y <- x + rnorm(10)
(rs <- rstat(x, y, units = bquote(m^-1)))
plot(x, y)
legend("bottomright", lstat(rs), bty = "n")

}
