% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm_semi_analytical.R
\name{rta_sa}
\alias{rta_sa}
\title{Radiative transfer semi-analytical approximations}
\usage{
rta_sa(a, bb, theta_s = 0, depth = Inf, rho_b, theta_v = 0,
  wsp = 0, aop = c("rrs", "rho"), model = c("Albert-Mobley03",
  "Lee98"), bbp)
}
\arguments{
\item{a}{Total absorption coefficient (1/m).}

\item{bb}{Total back-scattering coefficient (1/m).}

\item{theta_s}{Sun zenith angle refracted underwater (radians).}

\item{depth}{Bottom depth, positive downwards (m).}

\item{rho_b}{Bottom bi-hemispherical reflectance (unitless).}

\item{theta_v}{Observation nadir angle refracted underwater (radians).}

\item{wsp}{Wind speed (m/s). Only for model "Albert-Mobley03".}

\item{aop}{Desired reflectance. One of: 'rho' or 'rrs'. See details.}

\item{model}{Semi-analytical parametrization to use. Possible values are:
"Albert-Mobley03" and "Lee98". See details.}

\item{bbp}{Particle back-scattering (1/m). See details.}
}
\value{
A numeric vector with the subsurface bi-hemispherical ('r', 
unitless) or hemispherical-directional ('rrs', 1/sr) water-leaving 
reflectance.
}
\description{
This function calculates the subsurface bi-hemispherical and 
hemispherical-directional water-leaving reflectance over optically deep 
or optically shallow waters.
}
\details{
In Hydrology Optics, the bi-hemispherical and hemispherical-
-directional reflectances are commonly named irradiance refelctance ("R" or 
"Rho") and remote sensing reflectance ("Rrs"), respectively. Those are two 
apparent optical properties of the medium. This function uses semi-analytical 
parametrizations to calculate their subsurface values, i.e., at depth 0 just 
below the water surface. The lower case names are used to indicate the 
subsurface quantities. Those quantities can be propagated into air, just 
above the water surface, with the function \code{propagate_r}.

An optically deep medium is equivalent of a semi-infinite medium, i.e., the 
bottom depth is infinite and has no contribution to the water-leaving 
reflectance. Optically shallow water are defined as the complement, when 
there is "measurable" influence of bottom albedo (bi-hemispherical 
reflectance).

Model "Albert-Mobley03" is used as default. It implements the equations and 
coefficients provided in Albert & Mobley (2003). The model includes an 
extended range of water medium properties, the effect of wind speed and 
include specific coefficients for "rho" and "rrs".

Model "Lee98" is also available. It implements the equations and coefficients
provided in Lee et al. (1998), with the corrections of Lee et al. (1999). The 
model is specifically parametrized for "rrs", and "rho" is calculated (with a 
warning) by assuming that the diffuse subsurface hemispherical-directional 
water-leaving reflectance is Lambertian.

Both models can be used to calculate 'rrs' at off-nadir observation. The 
model "Lee98" requires that the backscattering of particles (bbp) be provided 
separately in case of off-nadir observation. The "Albert-Mobley03" also 
allows to calculate off-zenith illumination, while the model of "Lee98" is an
average fit to three Sun zenith angles (0, 30 and 60 degrees, in air).

If the arguments passed to the function are outside the range used for 
parametrization of the specific model, values will be return with a warning 
of extrapolation beyond model domain.
}
\examples{
theta_s  <- rad(seq(0, 90, 0.5))
tehta_sw <- snell(theta_s, 1, 1.33)
r_am     <- rta_sa(a = 1, bb = 0.5, theta_s = theta_sw, model = "Albert-Mobley03")
r_lee    <- rta_sa(a = 1, bb = 0.5, theta_s = theta_sw, model = "Lee98")
plot(deg(theta_s), r_am)
lines(deg(theta_s), r_lee)
}
\references{
Albert, A.; Mobley, C. D. 2003. An analytical model for subsurface irradiance 
and remote sensing reflectance in deep and shallow case-2 waters. Optics 
Express 11, 22, 2873-2890. DOI: 10.1364/oe.11.002873

Lee, Z.-P.; Carder, K. L.; Mobley, C. D.; Steward, R. G.; Patch, J. S. 1998. 
Hyperspectral remote sensing for shallow waters. I. A semianalytical model. 
Applied Optics 37, 27, 6329-6338. DOI: 10.1364/AO.37.006329

Lee, Z.-P.; Carder, K. L.; Mobley, C. D.; Steward, R. G.; Patch, J. S. 1999. 
Hyperspectral remote sensing for shallow waters: 2. Deriving bottom depths 
and water properties by optimization. Applied Optics 38, 18, 3831-3843. DOI:
10.1364/AO.38.003831
}
\seealso{
\code{\link{propagate_r}}
}
