% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/capturer.R
\name{SetupCapture}
\alias{SetupCapture}
\title{Setup information capturing for list of function

This function is respinsible for setting up capturing for functions}
\usage{
SetupCapture(flist, verbose = testrOptions("verbose"))
}
\arguments{
\item{flist}{function or list of functions to turn on capturing for. List should be only as character.}

\item{verbose}{if to print what functions will be captured}
}
\description{
Setup information capturing for list of function

This function is respinsible for setting up capturing for functions
}
\seealso{
Decorate
}

