% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specconv.R
\name{specconv}
\alias{specconv}
\title{Spectral convolution}
\usage{
specconv(spec, rsrf)
}
\arguments{
\item{spec}{A matrix containing spectral data, where the first column is the 
wavelength reference of each measurement.}

\item{rsrf}{A matrix containing the Relative Spectral Response Function of 
the sensor, where the first column is the wavelength reference.}
}
\value{
A matrix with number of rows equal to the number of bands and number 
of columns equal to number of spectral data in spec.
}
\description{
This function calculates the sum approximation for the convolution of a 
hyperspectral signal with a given sensor relative spectral response function.
}
\details{
The function calculates the sum approximation of the convolution. 
The RSRF is linearly interpolated to the spectral reference of the spec data, 
therefore spectral data beyond the interval of spec is implicitly taken as 
zero.
}
\seealso{
\code{rsrf}
}
