% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/testgen.r
\name{TestGen}
\alias{TestGen}
\title{Test Case generator based on capture files}
\usage{
TestGen(root, output.dir, use.get.anywhere = TRUE,
  verbose = testrOptions("verbose"))
}
\arguments{
\item{root}{a directory containg capture information or capture file}

\item{output.dir}{directory where generated test cases will be saved}

\item{use.get.anywhere}{whether to perform lookup for a closure in loaded namespaces}

\item{verbose}{wheater display debug output}
}
\description{
This function works with the trace information generated by instrumented GNU-R.
It is strictly oriented to that, please see readme for more information.
}

