% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iop_water.R
\name{vsf_water}
\alias{vsf_water}
\alias{b_water}
\title{Scattering of pure (saline) water}
\usage{
vsf_water(lambda, S = 0, Tc = 20, psi = pi/2, P = 0)

b_water(lambda, S = 0, Tc = 20, P = 0)
}
\arguments{
\item{lambda}{Wavelength in vacuum (nm).}

\item{S}{Salinity [0,40] (parts per thousand).}

\item{Tc}{Temperature [0,30] (ºC).}

\item{psi}{Polar scattering angle (radians).}

\item{P}{Pressure (bar in excess of 1 ATM).}
}
\value{
A numeric vector with the volume scattering function (m^-1 sr^-1) of 
pure (saline) water for function \code{vsf_water} and a numeric vector with 
the scattering coefficient.
}
\description{
This functions calculate the angular integral (scattering coefficient) and 
the angular distribution (volume scattering function) of  pure (saline) water 
scattering.
}
\details{
The formulation of scattering of pure (saline) water is based on the 
Einstein-Smoluchowski theory of scattering in a particle free media due to 
fluctuations of the dieletric constant (square of the refractive index) 
caused by random motion of molecules. This implementation follows Zhang & Hu
(2009A) that revisited the equations formulated with respect to the density 
derivative of the refractive index, after the new determinations of Proutiere 
et al. (1992). The excess scattering caused by diluted salts is based on salt 
concentration fluctuations by incorporating the salinity derivative of the 
refractive index (Zhang & Hu 2009B).

The density fluctuations are given by thermodynamic statistics and since both 
the density, the refractive index and the isothermal compressibility of water
are a function of pressure, the pressure is allowed to vary. Note however 
that the model is only validated for surface pressure. The default pressure 
therefore is 0 bar, with P defined as excess pressure to 1 ATM (in bar).

The volume scattering function is provided per polar scattering angle. In 
natural waters, the angular distribution of scattering is only a function of 
the polar angle of scattering, i.e, it is constant for all azimuths at a 
given polar scattering angle (Jonaz & Fournier, 2007).

Finally, this function will provide values slightly higher than the ones 
provided by the Water Optical Properties Processor (WOPP; Röttgers et al., 
2011) due to small differences in the physical constants, in particular the 
molecular weight of pure water. The required physical constants are taken 
from: \url{http://physics.nist.gov/cuu/Constants/}
}
\examples{
# Retrieve the volume scattering function of average seawater at 550 nm:
psi <- seq(0, pi, length.out = 180)
vsf_water(lambda = 550, S = 34.72, Tc = 3.5, psi = psi)

# Retrieve the scattering coefficient of average seawater in the visible 
range:
b_water(lambda = 400:700, S = 34.72, Tc = 3.5)

}
\seealso{
\code{\link{a_water}}, \code{\link{n_water}}
}
