% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_nir_sim}
\alias{water_nir_sim}
\title{NIR water similarity spectrum}
\format{An object of class \code{data.frame} with 101 rows and 4 columns.}
\usage{
water_nir_sim
}
\description{
This data set was compiled from high quality in situ observations as the 
average reflectance spectral shape of optically deep water targets from 
moderately turbid (SPM of 0.3 g/m3) to highly turbid (SPM of 200 g/m3) 
conditions. Data is normalized to 780 nm.
}
\references{
Ruddick, K. G.; Cauwer, V. De; Park, Y.-j.; Moore, G. 2006. Seaborne 
measurements of near infrared water-leaving reflectance: The similarity 
spectrum for turbid waters. Limnology and Oceanography 51, 2, 1167-1179. 
DOI: 10.4319/lo.2006.51.2.1167
}
\keyword{datasets}
