% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm_semi_analytical.R
\name{propagate_r}
\alias{propagate_r}
\title{Propagates reflectance across air-water interface}
\usage{
propagate_r(r, to = c("air", "water"), aop = c("rrs", "rho"))
}
\arguments{
\item{r}{Reflectance (unitless or 1/sr).}

\item{to}{One of: "air" or "water".}
}
\description{
This function propagates the reflectance across the air-water interface.
}
\details{
This function applies the average coefficients for water 
transmittance and refractive index of water relative to air to calculate the
scaling of the reflectance signal just above or just below the water surface.
The base equation and data is described in Gordon et al. (1988).

Irradiance reflectance just above the water surface (R(+)) is related to the 
irradiance reflectance just below the water surface (R(-)) by:

R(+) = t(Lu) * (na/nw)^2 * t(Ed) * R(-) / (1 - r(Eu) * R(-)),

where t(Lu) is the transmittance of in water upwelling radiance to air, 
(na/nw)^2 takes in account the radiance invariance law, t(Ed) in the 
transmittance of the in air downwelling irradiance to water and r(Eu) is the
reflectance of the in water upwelling radiance by the interface. A similar 
relation is found for Rrs(+) and rrs(-) by including the factor 1/Q, where Q
is the relation Eu/Lu. The interface transmittance and reflection 
coefficients depend on its roughness and the directional irradiance 
distribution (Gordon 2005) and on the wavelength dependent refractive indexes 
of air and water (Voss & Flora, 2017).
}
\references{
Voss, K. J.; Flora, S. 2017. Spectral dependence of the seawater-air radiance 
transmission coefficient. Journal of Atmospheric and Oceanic Technology 34, 
6, 1203-1205. DOI: 10.1175/JTECH-D-17-0040.1

Gordon, H. R.. 2005. Normalized water-leaving radiance: Revisiting the 
influence of surface roughness. Applied Optics, 44, 241-248. 
DOI: 10.1364/AO.44.000241

Gordon, H. R.; Brown, O. B.; Evans, R. H.; Brown, J. W.; Smith, R. C.; Baker, 
K. S.; Clark, D. K. 1988. A semianalytic radiance model ofocean color. 
Journal of Geophysical Research 93, 10, 909-10.924. DOI: 
10.1029/JD093iD09p10909

Lee, Z.-P.; Carder, K. L.; Mobley, C. D.; Steward, R. G.; Patch, J. S. 1998. 
Hyperspectral remote sensing for shallow waters. I. A semianalytical model. 
Applied Optics 37, 27, 6329-6338. DOI: 10.1364/AO.37.006329

Lee, Z.-P.; Carder, K. L.; Arnone, R. A. 2002. Deriving inherent optical 
properties from water color: a multiband quasi-analytical algorithm for 
optically deep waters. Applied Optics 41, 27, 5755-5772. DOI: 
10.1364/AO.41.005755
}
\seealso{
\code{\link{rta_sa}}
}
